Consolidate testing more.
