+++ src/lib/tests/ShardManagerTest.cpp
@@ -1,6 +1,3 @@
-#include <folly/init/Init.h>
-#include <folly/executors/GlobalExecutor.h>
-
 #include <homeobject/shard_manager.hpp>
 #include "lib/tests/fixture_app.hpp"
 
@@ -65,15 +62,4 @@ TEST_F(TestFixture, SealShard) {
             EXPECT_EQ(info.state, ShardInfo::State::SEALED);
         });
     }
-}
-
-int main(int argc, char* argv[]) {
-    int parsed_argc = argc;
-    ::testing::InitGoogleTest(&parsed_argc, argv);
-    SISL_OPTIONS_LOAD(parsed_argc, argv, logging, homeobject_options, test_home_object);
-    sisl::logging::SetLogger(std::string(argv[0]));
-    spdlog::set_pattern("[%D %T.%e] [%n] [%^%l%$] [%t] %v");
-    parsed_argc = 1;
-    auto f = ::folly::Init(&parsed_argc, &argv, true);
-    return RUN_ALL_TESTS();
 }
